extracted out updatePaths
