+++ src/renderer/components/configure-recording-set-page/index.tsx
@@ -3,30 +3,13 @@ import React, { FC, MouseEventHandler } from 'react';
 import { useLocalStorage, usePrevious } from 'react-use';
 
 import { Consumer, RecordingSet } from '../../types';
-import { checkFileExistence, getLSKey, join } from '../../utils';
+import { getLSKey } from '../../utils';
+import { checkPaths } from '../../utils/configure-recording-index-utils';
 
 import { ConfigureRecordingSet } from './configure-recording-set';
 import './show-details.scss';
 import './index.scss';
 import { PreviewPage } from './preview-page';
-
-const DELTA_ENGLISH_STATE_FILEPATH_MAP = {
-  'oto-ini': join('external', 'delta_eng_ver5', 'mkototemp.ini'),
-  'list-preview': join('external', 'delta_eng_ver5', 'デルタ式engver5_reclist.txt'),
-  dvcfg: join('external', 'To be added later'),
-};
-
-const CVVC_NORMAL_STATE_FILEPATH_MAP = {
-  'oto-ini': join('external', 'z.cvvc_normal', 'oto.ini'),
-  'list-preview': join('external', 'z.cvvc_normal', 'Reclist.txt'),
-  dvcfg: join('external', 'To be added later'),
-};
-
-const CUST_LIST_STATE_FILEPATH_MAP = {
-  'oto-ini': join('external', 'oto.ini'),
-  'list-preview': join('external', 'Reclist.txt'),
-  dvcfg: join('external', 'voice.dvcfg'),
-};
 
 /*
  Represents the page states controlled by configure-recording-set.
@@ -78,60 +61,16 @@ const ConfigureRecordingSetPage: FC<{
     '',
   );
 
-  const updatePaths = async (listName: string, isBuiltIn: boolean) => {
-    const states: MetadataState[] = [];
-    let isSubscribed = true;
-
-    const getFilePromise = async (
-      metaDataState: MetadataState,
-      listName: string,
-      isBuiltIn: boolean,
-      setFilePath: Consumer<string>,
-    ) => {
-      let filePath = '';
-      if (isBuiltIn) {
-        switch (listName) {
-          case 'デルタ式英語リストver5 (Delta English Ver. 5)':
-            filePath = join(filePath, DELTA_ENGLISH_STATE_FILEPATH_MAP[metaDataState]);
-            break;
-          case 'Z式CVVC-Normal (Z Chinese CVVC - Normal)':
-            filePath = join(filePath, CVVC_NORMAL_STATE_FILEPATH_MAP[metaDataState]);
-            break;
-        }
-      } else {
-        filePath = join(listName, CUST_LIST_STATE_FILEPATH_MAP[metaDataState]);
-      }
-
-      const response = checkFileExistence(filePath);
-      const fileExistenceAnswer = await response;
-      if (isSubscribed) {
-        if (fileExistenceAnswer == 'file') {
-          states.push(metaDataState);
-          setFilePath(filePath);
-        } else {
-          setFilePath('');
-        }
-      }
-    };
-    await Promise.all([
-      getFilePromise('oto-ini', listName, isBuiltIn, setOtoFilePath),
-      getFilePromise('list-preview', listName, isBuiltIn, setRecFilePath),
-      getFilePromise('dvcfg', listName, isBuiltIn, setDvcfgFilePath),
-    ]).catch((error) => {
-      log.error(error);
-      throw error;
-    });
-
-    if (states.length == 0) {
-      setMetadataStateIndex(-1);
-    } else {
-      setDropdownStateArray(states);
-      setMetadataStateIndex(0);
-    }
-
-    return () => {
-      isSubscribed = false;
-    };
+  const updatePaths = (listName: string, isBuiltIn: boolean) => {
+    return checkPaths(
+      listName,
+      isBuiltIn,
+      setOtoFilePath,
+      setRecFilePath,
+      setDvcfgFilePath,
+      setMetadataStateIndex,
+      setDropdownStateArray,
+    );
   };
   const currentState = dropDownStates[metadataStateIndex];
